requirements.txt
# requirements.txt
python-dotenv
pycryptodome
Pillow
